code/requirements.txt
python-dotenv==1.0.1
Telethon==1.35.0
